chatbot_parser/requirements.txt
requests==2.28.2
Flask==2.1.1
psycopg2-binary
